scraper/requirements.txt
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
exceptiongroup==1.3.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
h11==0.16.0
idna==3.11
oauthlib==3.3.1
outcome==1.3.0.post0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.2.0
